requirements.txt
reportlab>=4.2.0
Pillow>=10.0.0
colorthief>=0.2.1
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
